+++ src/duration.rs
@@ -1,9 +1,7 @@
 use core::str::FromStr;
 
 use crate::parsers;
-use crate::parsers::Stream;
 use alloc::string::String;
-use winnow::stream::StreamIsPartial;
 
 /// A time duration.
 ///
@@ -160,10 +158,7 @@ impl From<Duration> for ::core::time::Duration {
 /// let duration = winnow_iso8601::duration("P2015-11-03T21:56").unwrap();
 /// ```
 pub fn duration(string: &str) -> Result<Duration, String> {
-    let i = &mut Stream::new(string.as_bytes());
-    let _ = i.complete();
-
-    match parsers::parse_duration(i) {
+    match parsers::parse_duration(&mut string.as_bytes()) {
         Ok(p) => Ok(p),
         Err(e) => Err(format!("Failed to parse duration {}: {}", string, e)),
     }